requirements.txt
aiohttp==3.7.3
appdirs==1.4.3
async-timeout==3.0.1
attrs==20.3.0
aiocron==1.8
boto3==1.9.16
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
discord.py==1.7.3
distlib==0.3.0
distro==1.4.0
Flask==2.0.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
multidict==5.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2020.5
requests==2.22.0
retrying==1.3.3
sentry-sdk==1.1.0
six==1.14.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==2.0.1
yarl==1.6.3d
